requirements.txt
python-dotenv
google-generativeai
kafka-python
requests
confluent-kafka
Faker
orjson
avro